+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - **REFACTOR**: Update `ShadCalendar` `yearSelectorMinWidth` from `100` to `64` and `monthSelectorMinWidth` from `120` to `64`.
 - **REFACTOR**: Update `ShadCalendar` position of dropdown, which is now centered in the header.
+- **REFACTOR**: Add `dropdownFormatMonth` and `dropdownFormatYear` to `ShadCalendar`.
 - **FIX**: Center the placeholder in `ShadAvatar`.
 - **DOCS**: Fix missing `MainAxisSize.min` in `ShadCard` notification example.
 - **DOCS**: Adjust sheet docs example, with some spacing between buttons.